+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# v0.1.0 Elvish Ranger
+
+- Added external dependency on `rangy`
+- Update font size API to deal with `px` font sizes if `options.pixels` is `true`
+
 # v0.0.11 Botomless Pit
 
 - Added support for alignment